requirements.txt
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
googleapis-common-protos==1.58.0
grpcio==1.51.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
opentelemetry-api==1.15.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
protobuf==4.21.12
requests==2.28.2
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.13.0
